change type's datatype in Room entity

+++ HospitalManagementCore/Domain/Entities/Room.cs
@@ -10,7 +10,7 @@ namespace HospitalManagementCore.Domain.Entities
         public int Id { get; set; }
         public int Number { get; set; }
         public bool IsAvailable { get; set; }
-        public int Type { get; set; }
+        public byte Type { get; set; }
         public int BlockFloor { get; set; }
     }
 }
